Added backward compatability back to Python 2.2.
I removed some of the changes I was working on for Python3 (mostly
exception systax). The reason for doing this is that it's impossible to
maintain a single source file that supports both Python2 and Python3, so
Python3 is never going to happen in this code anyway.** This is one of
the most annoying things about Python3. Maybe I'll create a build script
that strips out the Python3 crap so I can at least maintain one set of
source files that can be retargeted to Python2 or Python3 as a
pre-install step.

** This may not be strictly true, but short of adding
conditional tests and exceptions to handle run-time errors
I can't think of a clean way to do this.