Finish canonical mode processing docs and tests
Linux is certainly the rotten egg in this respect, I could find
the code for sending '\a' for canonical mode processing in the
various kernel drivers within just a few minutes -- with the exception
of linux. After several hours of investigation of the kernel tty
driver, I still can't find where exactly '\a', '0x07', or 'BEL'
is sent.

We just trust that it is, verified on debian Linux. We also discover
that PC_MAX_CANON is not used, but rather, a hardcoded value inside
the kernel driver of 4096; the length of a ring buffer data structure,
also tested by the distance of head <-> tail of such ring buffer.