some more travis-ci accomidations
python3.2 and 3.4 on travis-ci are not flushing data to
child processes' stdin, presumably because CR is not yet
sent. tests are modified to always use sendline().

python3.2 on debian linux does not exhibit this same
behavior, so this is another hail mary.  if we can't get
travis-ci to agree, we might just scrap these tests all
together, it only confirms and examplifies the behavior
documented in send().