attempt one remaining error for py3.2 on travis-ci
we know that sendline() wouldn't work once we've reached the buffer,
but on travis-ci, this means stdin sent *still* has not been flushed
or received by the child process -- for python3.2 only. So we go
ahead and send backspace and sendline *prior* to testing echo output
and bel. I'm not entirely sure we will see BEL where we expect, only
travis-ci can say for sure. It is passing on debian linux py3.2,
though.