Added support for sockets.
From the maintainer:

    We changed the inheritance around - in earlier versions of
    pexpect, fdspawn inherited from spawn, even though spawn is
    actually the more specific case (it needs to know about the
    bizarreness of ptys). Now spawn and fdspawn both inherit from
    spawnbase. This makes a lot more sense, but it may well be
    causing what you see. I'd suggest you make a separate subclass
    for dealing with sockets.

Added a socket_pexpect module which creates a socket_spawn class.
The socket_spawn class overrides the read_nonblocking method of
the fdspawn class. The read_nonblocking method uses select to
implement the timeout and check if there is anything to read from
the socket. If there is something to read, the superclass
read_nonblocking is called.