Use PATH from env argument to pexpect.spawn
Currently pexpect.spawn takes an env argument along the lines of
subprocess.Popen but fails to use PATH from that env when finding
the binary to run.

The result when running a binary not on the PATH of the parent
process is pexpect will raise an exception, or worse, find a
different binary with the same name.

This change passes the env from spawn into the which function
so the correct PATH is searched and adds test coverage.