pxssh.login require either username or ssh_config
If pxssh.login is called with an ssh_config then do not require username since it can be supplied via the config and some programs do not want to manage the username at all since it would require parsing the ssh config.

This commit removes the independent username parameter when formatting cmd and instead appends the flag and the username to ssh_options if a username is provided.

If neither username nor ssh_config is provided raise a TypeError in line with the error raised natively if username was not provided as a required positional argument.

Tests ensure that ssh_config lists a username for the server in question
so the user running the program does not leak. Note that ssh_config
keywords are case-insensitive.

This commit by itself fails tests because it has been modified to
simplifiy the merge commit.