pxssh.login: allow using implicit SSH configuration
With OpenSSH clients, configuration data can come from multiple
sources (see man ssh_config):
1.   command-line options
2.   user's configuration file (~/.ssh/config)
3.   system-wide configuration file (/etc/ssh/ssh_config)

This means for example that the configuration for a particular host
could come from a user-defined configuration file, while for another
host it comes from the system-wide configuration file.

Besides, pexpect forces the use of either a username, or a specific
configuration file that will be parsed and checked for the host (among
other things).

As a consequence, it's currently not possible to use pexpect like a
regular OpenSSH client would be used, with multiple hosts potentially
defined in different configuration files.

To address this issue, instead of forcing the use of either a username
or a configuration file, just let the ssh client handle it. Also
remove the associated test.

This also allows the configuration files that are implicitly used to
contain more complex directives that are not currently supported when
parsing `ssh_config` (e.g. the `Include` directive).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>