Do not send more characters than can be buffered by cat
The cat command on FreeBSD blocks if the default input buffer fills
and not output can be sent. The test used to generate more than 10 KB
of data before accepting the output generated by the cat program,
causing cat to block (stop reading input) until the output buffer has
been drained.

There is no reason to send long lines in this test, therefore it is
easily possible to perform the tests with lines that do not cause more
than 4 KB (the default input buffer size on FreeBSD) to be sent before
output from the cat program is checked.

In order to not depend on such buffer sizes pexpect needs to check for
output from the spawned command and buffer it, before it gets to
execute the expect() function. Apparently this is not happening and
thus the system dependent buffer size limit makes this test fail if
its capacity is exceeded.