Test cases for specific shell behaviour
Adds support to fakessh to emulate a specific shell with respect to
prompt setting. The shell is specified as a parameter after the
hostname, similar to real ssh taking a command to execute on the remote.
This is slightly awkward when calling pxssh.login() as it means giving
the server parameter as 'hostname shell', but this is only needed for
the unit tests.