Set prompt correctly for zsh
When the remote system/user is using zsh as their login shell, `\$` is
not an escape sequence for the prompt, and so self.PROMPT fails to match
because the backslash is included in the received text. The equivalent
escape sequence for zsh is `%(!.#.$)`, so add a self.PROMPT_SET_ZSH using
this sequence.

Fixes #711